Refactor mcp-server.js: Extract modules and reduce file size by 82%
- Extracted metrics and logging to mcp-server-metrics.js
- Extracted tool merging logic to mcp-tool-merger.js
- Extracted notification management to mcp-notification-manager.js
- Extracted server lifecycle to mcp-server-lifecycle.js
- Removed all deprecated Legacy methods
- Removed deprecated wrapper methods that just delegate
- Simplified notifyRouteChanges to use tool merger
- File reduced from 2,225 lines to 402 lines (82% reduction)
- Improved code organization and maintainability
- All tests passing
